scale controlling added
