+++ js/upload-image.js
@@ -1,7 +1,7 @@
 import {disableEscHandling} from './util.js';
 import {createUploadImageHandler} from './upload-img-listeners.js';
 import {createValidation} from './upload-validation.js';
-import {SCALE_IMAGE_MAX, SCALE_IMAGE_MIN, SCALE_IMAGE_STEP, SCALE_IMAGE_DEFAULT} from './data.js';
+import {createScaleControlling} from './upload-scale-contol.js';
 
 const uploadForm = document.querySelector('.img-upload__form');
 const uploadImageInput = uploadForm.querySelector('.img-upload__input');
@@ -9,6 +9,11 @@ const uploadOverlay = uploadForm.querySelector('.img-upload__overlay');
 const cancelUploadButton = uploadForm.querySelector('.img-upload__cancel');
 const hashTagInput = uploadForm.querySelector('.text__hashtags');
 const commentField = uploadForm.querySelector('.text__description');
+const scaleContainer = uploadForm.querySelector('.img-upload__scale');
+const decreaseScaleButton = scaleContainer.querySelector('.scale__control--smaller');
+const increaseScaleButton = scaleContainer.querySelector('.scale__control--bigger');
+const scaleControlValue = scaleContainer.querySelector('.scale__control--value');
+const imageToUpload = uploadForm.querySelector('.img-upload__preview img');
 
 disableEscHandling(hashTagInput);
 disableEscHandling(commentField);
@@ -41,35 +46,7 @@ uploadForm.addEventListener('submit', (evt) => {
   // }
 });
 
-// кнопка масштаба
-
-const scaleContainer = document.querySelector('.img-upload__scale');
-const decreaseScaleButton = scaleContainer.querySelector('.scale__control--smaller');
-const increaseScaleButton = scaleContainer.querySelector('.scale__control--bigger');
-const scaleControlValue = scaleContainer.querySelector('.scale__control--value');
-const imageToUpload = uploadForm.querySelector('.img-upload__preview img');
-
-scaleControlValue.value = `${SCALE_IMAGE_DEFAULT}%`;
-let currentValue = parseInt(scaleControlValue.value, 10);
-
-const decreaseValue = () => {
-  if (currentValue <= SCALE_IMAGE_DEFAULT && currentValue > SCALE_IMAGE_MIN) {
-    currentValue -= SCALE_IMAGE_STEP;
-    scaleControlValue.value = `${currentValue}%`;
-    imageToUpload.style.transform = `scale(${currentValue / 100})`;
-  }
-};
-
-const increaseValue = () => {
-  if (currentValue >= SCALE_IMAGE_MIN && currentValue < SCALE_IMAGE_MAX) {
-    currentValue += SCALE_IMAGE_STEP;
-    scaleControlValue.value = `${currentValue}%`;
-    imageToUpload.style.transform = `scale(${currentValue / 100})`;
-  }
-};
-
-decreaseScaleButton.addEventListener('click', decreaseValue);
-increaseScaleButton.addEventListener('click', increaseValue);
+createScaleControlling(scaleControlValue, imageToUpload, decreaseScaleButton, increaseScaleButton);
 
 //Управление эффектами
